requirements.txt
# Web Framework
flask>=2.0.0
flask-cors>=4.0.0

# OpenAI and Related
openai>=1.3.0
instructor>=0.4.8
langsmith>=0.0.69
tiktoken>=0.5.2

# Vector Database
pinecone-client>=3.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
firecrawl>=0.1.0

# Machine Learning
transformers>=4.36.0
torch>=2.1.0
pydantic>=2.5.0

# Utilities
python-dotenv==1.0.0
cryptography>=41.0.0
termcolor>=2.4.0
tenacity>=8.2.0

# Development Tools
black>=23.0.0  # Code formatting
pytest>=7.4.0  # Testing
mypy>=1.7.0   # Type checking

# Additional Dependencies
aiohttp>=3.9.0  # For async HTTP requests
pandas>=2.1.0   # For data manipulation
numpy>=1.24.0   # For numerical operations
